refactor: use subfolders for @test-utils imports
