+++ src/test-utils/create-use-rtk-query-mock-return-value.ts
@@ -1,0 +1,15 @@
+import { vi } from 'vitest'
+import { UseRtkQueryResult } from '@hooks'
+
+export const createUseRtkQueryMockReturnValue = (
+    payload: Partial<UseRtkQueryResult>
+): UseRtkQueryResult => {
+    return {
+        data: {},
+        refetch: vi.fn(),
+        isLoading: false,
+        isError: false,
+        error: undefined,
+        ...payload,
+    } as UseRtkQueryResult
+}
