--- src/test-utils/index.tsx
@@ -1,47 +1,0 @@
-import { configureStore, ReducersMapObject, Store } from '@reduxjs/toolkit'
-import { render } from '@testing-library/react'
-import { PropsWithChildren } from 'react'
-import type { ReactElement } from 'react'
-import { Provider } from 'react-redux'
-import { vi } from 'vitest'
-import { UseRtkQueryResult } from '@hooks'
-import type { RootState } from '@types'
-
-export const setupStore = (
-    reducer: Partial<ReducersMapObject<RootState>>,
-    preloadedState: Partial<RootState>
-): {
-    getState: () => Partial<RootState>
-    dispatch: ReturnType<typeof configureStore>['dispatch']
-    subscribe: ReturnType<typeof configureStore>['subscribe']
-    replaceReducer: ReturnType<typeof configureStore>['replaceReducer']
-} => {
-    return configureStore({
-        reducer: reducer as ReducersMapObject<RootState>,
-        preloadedState,
-    })
-}
-
-export const renderWithReduxStoreProvider = (
-    ui: ReactElement,
-    store: ReturnType<typeof setupStore>
-) => {
-    const Wrapper = ({ children }: PropsWithChildren) => (
-        <Provider store={store as Store}>{children}</Provider>
-    )
-
-    return render(ui, { wrapper: Wrapper })
-}
-
-export const createUseRtkQueryMockReturnValue = (
-    payload: Partial<UseRtkQueryResult>
-): UseRtkQueryResult => {
-    return {
-        data: {},
-        refetch: vi.fn(),
-        isLoading: false,
-        isError: false,
-        error: undefined,
-        ...payload,
-    } as UseRtkQueryResult
-}
